refactor: move `GroupOptions` into common
