+++ google/cloud/google_cloud_cpp_common_unit_tests.bzl
@@ -34,6 +34,7 @@ google_cloud_cpp_common_unit_tests = [
     "internal/filesystem_test.cc",
     "internal/format_time_point_test.cc",
     "internal/future_impl_test.cc",
+    "internal/group_options_test.cc",
     "internal/invoke_result_test.cc",
     "internal/log_impl_test.cc",
     "internal/pagination_range_test.cc",
