+++ google/cloud/internal/group_options_test.cc
@@ -12,33 +12,31 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "google/cloud/storage/internal/group_options.h"
-#include "google/cloud/storage/well_known_parameters.h"
+#include "google/cloud/internal/group_options.h"
 #include "google/cloud/common_options.h"
 #include <gmock/gmock.h>
 
 namespace google {
 namespace cloud {
-namespace storage {
 GOOGLE_CLOUD_CPP_INLINE_NAMESPACE_BEGIN
 namespace internal {
 namespace {
 
-google::cloud::Options SimulateRawClientOptions() {
+Options TestOptions() {
   return Options{}
       .set<UserProjectOption>("u-p-default")
       .set<AuthorityOption>("a-default");
 }
 
-TEST(MakeOptionSpanTest, JustDefaults) {
-  auto const group = GroupOptions(SimulateRawClientOptions());
+TEST(MakeOptionSpanTest, JustOptions) {
+  auto const group = GroupOptions(TestOptions());
   EXPECT_EQ("u-p-default", group.get<UserProjectOption>());
   EXPECT_EQ("a-default", group.get<AuthorityOption>());
 }
 
 TEST(MakeOptionSpanTest, Overrides) {
   auto const group =
-      GroupOptions(SimulateRawClientOptions(),
+      GroupOptions(TestOptions(),
                    Options{}
                        .set<EndpointOption>("test-endpoint")
                        .set<AuthorityOption>("a-override-1"),
@@ -49,28 +47,31 @@ TEST(MakeOptionSpanTest, Overrides) {
 }
 
 TEST(MakeOptionSpanTest, OverridesMixedWithRequestOptions) {
+  int x = 5;
+  int const& y = x;
+  int const z = 10;
+  struct Thing {};
+
   auto const group = GroupOptions(
-      SimulateRawClientOptions(), IfGenerationMatch(0),
-      Options{}.set<EndpointOption>("test-endpoint"), IfGenerationNotMatch(0),
-      Options{}.set<AuthorityOption>("a-override-1"), IfMetagenerationMatch(0),
-      Options{}.set<AuthorityOption>("a-override-2"),
-      IfMetagenerationNotMatch(0), Generation(7), IfGenerationMatch(0));
+      "string", TestOptions(), Thing{},
+      Options{}.set<EndpointOption>("test-endpoint"), 5,
+      Options{}.set<AuthorityOption>("a-override-1"), x,
+      Options{}.set<AuthorityOption>("a-override-2"), y, std::move(z));
   EXPECT_EQ("u-p-default", group.get<UserProjectOption>());
   EXPECT_EQ("a-override-2", group.get<AuthorityOption>());
   EXPECT_EQ("test-endpoint", group.get<EndpointOption>());
 }
 
 TEST(MakeOptionSpanTest, Declaration) {
-  auto const g1 = GroupOptions(SimulateRawClientOptions(), 5);
+  auto const g1 = GroupOptions(TestOptions(), 5);
   EXPECT_EQ("u-p-default", g1.get<UserProjectOption>());
 
-  auto const g2 = GroupOptions(5, SimulateRawClientOptions());
+  auto const g2 = GroupOptions(5, TestOptions());
   EXPECT_EQ("u-p-default", g2.get<UserProjectOption>());
 }
 
 }  // namespace
 }  // namespace internal
 GOOGLE_CLOUD_CPP_INLINE_NAMESPACE_END
-}  // namespace storage
 }  // namespace cloud
 }  // namespace google
